Added System Tab and removed in/out cluster deployment tab
