+++ ui/components/MeshConfigComponent.js
@@ -1,149 +1,344 @@
-// @ts-check
-import React from "react";
-import PropTypes from "prop-types";
-import { withStyles } from "@material-ui/core/styles";
 import {
-  NoSsr,
-  Button,
-  FormGroup,
-  InputAdornment,
-  Chip,
-  IconButton,
-  MenuItem,
-  Tooltip,
-  Paper,
-  Grid,
-  FormControlLabel,
-  Switch,
-  TextField,
-  List,
-  ListItem,
-  ListItemText,
-  CircularProgress,
-  Select,
-  Box,
-  FormControl,
-  InputLabel,
-} from "@material-ui/core";
-import blue from "@material-ui/core/colors/blue";
-import CloudUploadIcon from "@material-ui/icons/CloudUpload";
-import { connect } from "react-redux";
-import { bindActionCreators } from "redux";
-import { withRouter } from "next/router";
-import { withSnackbar } from "notistack";
+  Chip, Grid, IconButton, List, ListItem, ListItemText, Menu, MenuItem, Switch,
+  Tooltip, Paper, NoSsr, TableCell, TableContainer, Table, Button, Typography,
+  TextField, FormGroup, InputAdornment
+} from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
 import CloseIcon from "@material-ui/icons/Close";
-import { updateK8SConfig, updateProgress } from "../lib/store";
-import dataFetch, { promisifiedDataFetch } from "../lib/data-fetch";
-import subscribeOperatorStatusEvents from "./graphql/subscriptions/OperatorStatusSubscription";
-import subscribeMeshSyncStatusEvents from "./graphql/subscriptions/MeshSyncStatusSubscription";
-import changeOperatorState from "./graphql/mutations/OperatorStatusMutation";
-import fetchMesheryOperatorStatus from "./graphql/queries/OperatorStatusQuery";
-import NatsStatusQuery from "./graphql/queries/NatsStatusQuery";
-import MeshsyncStatusQuery from "./graphql/queries/MeshsyncStatusQuery";
-import resetDatabase from "./graphql/queries/ResetDatabaseQuery";
-import PromptComponent from "./PromptComponent";
+import { withSnackbar } from "notistack";
+import { useState, useEffect, useRef } from 'react';
+import DataTable from "mui-datatables";
+import { withStyles } from '@material-ui/styles';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import { updateProgress } from '../lib/store';
+import subscribeOperatorStatusEvents from './graphql/subscriptions/OperatorStatusSubscription';
+import subscribeMeshSyncStatusEvents from './graphql/subscriptions/MeshSyncStatusSubscription';
+import dataFetch, { promisifiedDataFetch } from '../lib/data-fetch';
+import fetchMesheryOperatorStatus from './graphql/queries/OperatorStatusQuery';
+import PromptComponent from './PromptComponent';
+import CloudUploadIcon from "@material-ui/icons/CloudUpload";
+import MeshsyncStatusQuery from './graphql/queries/MeshsyncStatusQuery';
+import NatsStatusQuery from './graphql/queries/NatsStatusQuery';
+import changeOperatorState from './graphql/mutations/OperatorStatusMutation';
 
 const styles = (theme) => ({
-  clusterConfiguratorWrapper : { padding : theme.spacing(5), },
-  buttons : { display : "flex",
-    justifyContent : "flex-end", },
-  button : { marginTop : theme.spacing(3),
-    marginLeft : theme.spacing(1), },
-  buttonsCluster : { display : "flex",
-    justifyContent : "center", },
-  margin : { margin : theme.spacing(1), },
-  alreadyConfigured : { textAlign : "center",
-    padding : theme.spacing(20), },
-  colorSwitchBase : { color : blue[300],
-    "&$colorChecked" : { color : blue[500],
-      "& + $colorBar" : { backgroundColor : blue[500], }, }, },
-  colorBar : {},
-  colorChecked : {},
-  fileLabel : { width : "100%", },
-  fileLabelText : { cursor : "pointer",
-    "& *" : { cursor : "pointer", }, },
-  inClusterLabel : { paddingRight : theme.spacing(2), },
-  alignCenter : { textAlign : "center", },
-  alignLeft : { textAlign : "left",
-    marginBottom : theme.spacing(2), },
-  fileInputStyle : { display : "none", },
-  icon : { width : theme.spacing(2.5), },
-  configure : {
-    display : "inline-block",
-    width : "48%",
-    wordWrap : "break-word",
-    [theme.breakpoints.down(945)] : { width : "100%", },
-  },
-  vertical : {
-    display : "inline-block",
-    height : 150,
-    marginBottom : -60,
-    [theme.breakpoints.down(945)] : { display : "none", },
-  },
-  horizontal : { display : "none",
-    [theme.breakpoints.down(945)] : { display : "block", }, },
-  formconfig : {
-    display : "inline-block",
-    marginLeft : 30,
-    [theme.breakpoints.up(946)] : { width : "45%", },
-    [theme.breakpoints.down(945)] : { width : "100%",
-      marginLeft : 0, },
-  },
-  currentConfigHeading : {
-    display : "inline-block",
-    width : "48%",
-    textAlign : "center",
-    [theme.breakpoints.down(945)] : { width : "100%", },
-  },
-  changeConfigHeading : {
-    display : "inline-block",
-    width : "48%",
-    textAlign : "center",
-    [theme.breakpoints.down(945)] : { display : "none", },
-  },
   operationButton : {
     [theme.breakpoints.down(1180)] : {
       marginRight : "25px",
     },
   },
-  contentContainer : {
-    [theme.breakpoints.down(1050)] : {
-      flexDirection : "column",
-    },
-  },
+  icon : { width : theme.spacing(2.5), },
+  paper : { padding : theme.spacing(2), },
+  heading : { textAlign : "center", },
   configBoxContainer : {
     [theme.breakpoints.down(1050)] : {
       flexGrow : 0,
       maxWidth : '100%',
       flexBasis : '100%',
     },
+    [theme.breakpoints.down(1050)] : {
+      flexDirection : "column",
+    },
   },
-  changeConfigHeadingOne : { display : "none",
-    [theme.breakpoints.down(945)] : { display : "inline-block",
-      width : "100%",
-      textAlign : "center", }, },
-  buttonconfig : { display : "inline-block",
-    width : "48%",
-    [theme.breakpoints.down(945)] : { width : "100%", }, },
-  paper : { padding : theme.spacing(2), },
-  heading : { textAlign : "center", },
+  clusterConfiguratorWrapper : { padding : theme.spacing(5), display : "flex" },
+  contentContainer : {
+    [theme.breakpoints.down(1050)] : {
+      flexDirection : "column",
+    },
+    flexWrap : "noWrap",
+  },
+  paper : { margin : theme.spacing(2), },
+  fileInputStyle : { display : "none", },
+  button : { marginTop : theme.spacing(3),
+    marginLeft : theme.spacing(1), },
   grey : { background : "WhiteSmoke",
     padding : theme.spacing(2),
     borderRadius : "inherit", },
+  fileLabelText : { cursor : "pointer",
+    "& *" : { cursor : "pointer", }, },
+  subtitle : {
+    minWidth : 400,
+    overflowWrap : 'anywhere',
+    textAlign : 'left',
+    padding : '5px'
+  },
+  text : {
+    width : "80%",
+    wordWrap : "break-word"
+  }
 });
 
-async function fetchAllContexts(number) {
-  return await promisifiedDataFetch("/api/system/kubernetes/contexts?pageSize=" + number)
-}
+function MesherySettingsNew({ classes, enqueueSnackbar, closeSnackbar, updateProgress }) {
+  const [data, setData] = useState([])
+  let k8sfileElementVal ="";
+  const [showMenu, setShowMenu] = useState([false])
+  const [anchorEl, setAnchorEl] = useState(null);
+  const [operatorInstalled, setOperatorInstalled] = useState(false);
+  const [meshSyncInstalled, setMeshSyncInstalled] = useState(false);
+  const [meshSyncState, setMeshSyncState] = useState("N/A");
+  const [NATSState, setNATSState] = useState("UNKNOWN");
+  const [NATSVersion, setNATSVersion] = useState("N/A");
+  const [meshSyncVersion, setMeshSyncVersion] = useState("N/A");
+  const [operatorVersion, setOperatorVersion] = useState("N/A");
+  const [operatorProcessing, setOperatorProcessing] = useState(false);
+  const [operatorSwitch, setOperatorSwitch] = useState(false);
+  const [meshSyncStatusSubscription, setMeshSyncStatusSubscription] = useState(null);
+  const [operatorStatusSubscription, setOperatorStatusSubscription] = useState(null);
+  const [contexts, setContexts] = useState([]);
+  const [k8sVersion, setK8sVersion] = useState("N/A");
+  // const [toUploadContexts, setToUploadContexts] = useState([]);
+  // const [open, setOpen] = useState(false);
+  const [formData, setFormData] = useState(null);
 
-async function uploadK8sConfig() {
-  const field = document.getElementById("k8sfile");
-  if (field instanceof HTMLInputElement) {
-    if (field.files.length < 1) return;
-    const name = field.files[0].name;
-    const formData = new FormData();
-    formData.append("k8sfile", field.files[0])
+  const ref = useRef(null);
 
+  const dateOptions = { weekday : 'long', year : 'numeric', month : 'long', day : 'numeric' };
+
+  useEffect(() => {
+    let meshSyncStatusEventsSubscription = subscribeMeshSyncStatusEvents((res) => {
+      if (res.meshsync?.error) {
+        handleError(res.meshsync?.error?.description || "MeshSync could not be reached");
+        return;
+      }
+    });
+
+    let tableInfo = [];
+    fetchAllContexts(25)
+      .then(res => {
+        console.log(res, "CTX");
+        handleContexts(res.contexts);
+        res.contexts.forEach((ctx) => {
+          let data = {
+            context : ctx.name,
+            location : ctx.server,
+            deployment_type : "fix",
+            last_discovery : "",
+            name : ctx.name,
+            id : ctx.id
+          };
+          tableInfo.push(data);
+        })
+        setData(tableInfo);
+      })
+      .catch(handleError("failed to fetch contexts for the instance"))
+
+    getKubernetesVersion();
+
+    let operatorStatusEventsSubscription = subscribeOperatorStatusEvents(setOperatorState);
+    setOperatorStatusSubscription(operatorStatusEventsSubscription);
+    fetchMesheryOperatorStatus().subscribe({
+      next : (res) => {
+        setOperatorState(res);
+      },
+      error : (err) => console.log("error at operator scan: " + err),
+    });
+
+    setMeshSyncStatusSubscription(meshSyncStatusEventsSubscription);
+
+    return () => {
+      if (meshSyncStatusSubscription) {
+        meshSyncStatusSubscription.dispose();
+      }
+      if (operatorStatusSubscription) {
+        operatorStatusSubscription.dispose();
+      }
+    }
+  }, [])
+
+  const handleContexts = (contexts) => {
+    contexts.forEach((ctx) => {
+      let cdt = new Date(ctx.created_at);
+      let updt = new Date(ctx.updated_at);
+
+      ctx.created_at = cdt.toLocaleDateString("en-US", options)
+      + " " +  cdt.toLocaleTimeString("en-US");
+
+      ctx.updated_at = updt.toLocaleDateString("en-US", options)
+      + " " +  updt.toLocaleTimeString("en-US")
+    })
+    setContexts(contexts);
+  }
+
+  const handleMenuClose = (index) => {
+    let menu = [...showMenu];
+    menu[index] = false;
+    setShowMenu(menu)
+  }
+
+  const setOperatorState = (res) => {
+    console.log("incoming change")
+    if (res.operator?.error) {
+      handleError("Operator could not be reached")(res.operator?.error?.description);
+      setOperatorProcessing(false);
+      return false;
+    }
+
+    if (res.operator?.status === "ENABLED") {
+      setOperatorProcessing(false);
+      res.operator?.controllers?.forEach((controller) => {
+        if (controller.name === "broker" && controller.status.includes("CONNECTED")) {
+          setNATSState(controller.status);
+          setNATSVersion(controller.version);
+        } else if (controller.name === "meshsync" && controller.status.includes("ENABLED")) {
+          setMeshSyncInstalled(true);
+          setMeshSyncVersion(controller.version);
+          setMeshSyncState(controller.status);
+        }
+      });
+      setOperatorInstalled(true);
+      setOperatorSwitch(true);
+      setOperatorVersion(res.operator?.version);
+      return true;
+    }
+    if (res.operator?.status === "DISABLED") {
+      setOperatorProcessing(false);
+    }
+
+    if (res.operator?.status === "PROCESSING") {
+      console.log("setting to processing");
+      setOperatorProcessing(true);
+    }
+  }
+
+  async function fetchAllContexts(number) {
+    return await promisifiedDataFetch("/api/system/kubernetes/contexts?pageSize=" + number)
+  }
+
+  const handleError = (msg) => (error) => {
+    updateProgress({ showProgress : false });
+    enqueueSnackbar(`${msg}: ${error}`, { variant : "error", preventDuplicate : true,
+      action : (key) => (
+        <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
+          <CloseIcon />
+        </IconButton>
+      ),
+      autoHideDuration : 7000, });
+  };
+
+  const handleMenuOpen = (e, index) => {
+    setAnchorEl(e.currentTarget)
+    let menu = [...showMenu];
+    menu[index] = true;
+    setShowMenu(menu);
+    console.log(contexts, "CTX");
+  }
+
+  const handleSuccess = msg => {
+    updateProgress({ showProgress : false });
+    enqueueSnackbar(msg, {
+      variant : "success",
+      action : (key) => (
+        <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
+          <CloseIcon />
+        </IconButton>
+      ),
+      autoHideDuration : 7000,
+    });
+  }
+
+  const handleLastDiscover = (index) => {
+    let dt = new Date();
+    const newDate = dt.toLocaleDateString("en-US", dateOptions) + "  " + dt.toLocaleTimeString("en-US");
+    let newData = [...data];
+    newData[index].last_discovery = newDate;
+    setData(newData);
+  }
+
+  const getKubernetesVersion = () => {
+    dataFetch(
+      "/api/oam/workload/APIService.K8s",
+      { credentials : "same-origin" },
+      (result) => {
+        if (result) {
+          let version = result[0].oam_definition?.spec?.metadata?.version;
+          setK8sVersion(version);
+        }
+      }
+    )
+  }
+
+  const handleKubernetesClick = (context, index) => {
+    updateProgress({ showProgress : true });
+    dataFetch(
+      "/api/system/kubernetes/ping?context=" + context,
+      { credentials : "same-origin" },
+      (result) => {
+        updateProgress({ showProgress : false });
+        if (typeof result !== "undefined") {
+          handleLastDiscover(index);
+          enqueueSnackbar("Kubernetes was successfully pinged!", {
+            variant : "success",
+            "data-cy" : "k8sSuccessSnackbar",
+            autoHideDuration : 2000,
+            action : (key) => (
+              <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
+                <CloseIcon />
+              </IconButton>
+            ),
+          });
+        }
+      },
+      handleError("Kubernetes config could not be validated")
+    );
+  };
+
+  const handleOperatorSwitch = () => {
+    const variables = { status : `${!operatorSwitch
+      ? "ENABLED"
+      : "DISABLED"}`, };
+    updateProgress({ showProgress : true });
+
+    changeOperatorState((response, errors) => {
+      updateProgress({ showProgress : false });
+      if (errors !== undefined) {
+        handleError("Unable to install operator");
+      }
+      enqueueSnackbar("Operator " + response.operatorStatus.toLowerCase(), { variant : "success",
+        autoHideDuration : 2000,
+        action : (key) => (
+          <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
+            <CloseIcon />
+          </IconButton>
+        ), });
+      setOperatorSwitch(operatorSwitch => !operatorSwitch);
+    }, variables);
+  };
+
+  const handleConfigDelete = (id, index) => {
+    console.log("test");
+    updateProgress({ showProgress : true });
+    dataFetch(
+      "/api/system/kubernetes/contexts/" + id,
+      { credentials : "same-origin",
+        method : "DELETE" },
+      (result) => {
+        updateProgress({ showProgress : false });
+        if (index != undefined) {
+          let newData = data.filter((dt, idx) => index != idx);
+          console.log(index, "LL");
+          setData(newData);
+          console.log("result", result);
+        }
+      },
+      handleError("failed to delete kubernetes context")
+    );
+  }
+
+  const handleChange = () => {
+    const field = document.getElementById("k8sfile");
+    const textField = document.getElementById("k8sfileLabelText");
+    if (field instanceof HTMLInputElement) {
+      if (field.files.length < 1) return;
+      const name = field.files[0].name;
+      const formData = new FormData();
+      formData.append("k8sfile", field.files[0])
+      textField.value=name;
+      setFormData(formData);
+    }
+  }
+
+  const uploadK8SConfig = async () => {
     await promisifiedDataFetch(
       "/api/system/kubernetes",
       {
@@ -151,389 +346,459 @@ async function uploadK8sConfig() {
         body : formData,
       }
     )
-    return name;
+    console.log("changed");
   }
-}
-
-async function changeContext(id) {
-  return await promisifiedDataFetch("/api/system/kubernetes/contexts/current/" + id, { method : "POST" })
-}
-
-class MeshConfigComponent extends React.Component {
-  constructor(props) {
-    super(props);
-    const {
-      inClusterConfig, contextName, clusterConfigured, k8sfile, configuredServer
-    } = props;
-    this.state = {
-      inClusterConfig, // read from store
-      inClusterConfigForm : inClusterConfig,
-      k8sfile, // read from store
-      k8sfileElementVal : "",
-      contextName, // read from store
-      contextNameForForm : "",
-      contextsFromFile : [],
-      clusterConfigured, // read from store
-      configuredServer,
-      k8sfileError : false,
-      ts : new Date(),
-      contexts : [],
-      selectContext : contextName,
-
-      operatorInstalled : false,
-      operatorVersion : "N/A",
-      meshSyncInstalled : false,
-      meshSyncVersion : "N/A",
-      meshSyncState : "N/A",
-      NATSState : "UNKNOWN",
-      NATSVersion : "N/A",
-
-      operatorSwitch : false,
-      operatorProcessing : false,
-
-      fileName : "",
-      meshSyncStatusEventsSubscription : null,
-      operatorStatusEventsSubscription : null,
-
-    };
-    this.ref = React.createRef();
-  }
-
-  static getDerivedStateFromProps(props, state) {
-    const {
-      inClusterConfig, contextName, clusterConfigured, k8sfile, configuredServer
-    } = props;
-    if (props.ts > state.ts) {
-      let newState = {
-        inClusterConfig,
-        k8sfile,
-        k8sfileElementVal : "",
-        contextName,
-        clusterConfigured,
-        configuredServer,
-        ts : props.ts,
-      };
-
-      // If contextsFromFile is empty then add the default value to it
-      if (!state.contextsFromFile?.length)
-        newState = { ...newState, contextsFromFile : [{ contextName, currentContext : true }] };
-      return newState;
-    }
-    return {};
-  }
-
-
-  componentDidMount() {
-    const self = this;
-    // Subscribe to the operator events
-    let meshSyncStatusEventsSubscription = subscribeMeshSyncStatusEvents((res) => {
-      if (res.meshsync?.error) {
-        self.handleError(res.meshsync?.error?.description || "MeshSync could not be reached");
-        return;
-      }
-    });
-
-    fetchAllContexts(25)
-      .then(res => this.setState({ contexts : res.contexts }))
-      .catch(this.handleError("failed to fetch contexts for the instance"))
-
-    let operatorStatusEventsSubscription = subscribeOperatorStatusEvents(self.setOperatorState);
-    fetchMesheryOperatorStatus().subscribe({
-      next : (res) => {
-        self.setOperatorState(res);
-      },
-      error : (err) => console.log("error at operator scan: " + err),
-    });
-
-    self.setState({ meshSyncStatusEventsSubscription, operatorStatusEventsSubscription })
-  }
-
-
-  componentWillUnmount () {
-    this.state.meshSyncStatusEventsSubscription.dispose()
-    this.state.operatorStatusEventsSubscription.dispose()
-  }
-
-  setOperatorState = (res) => {
-    console.log("incoming change")
-    const self = this;
-    if (res.operator?.error) {
-      self.handleError("Operator could not be reached")(res.operator?.error?.description);
-      self.setState({ operatorProcessing : false })
-      return false;
-    }
-
-    if (res.operator?.status === "ENABLED") {
-      self.setState({ operatorProcessing : false })
-      res.operator?.controllers?.forEach((controller) => {
-        if (controller.name === "broker" && controller.status.includes("CONNECTED")) {
-          self.setState({
-            NATSState : controller.status,
-            NATSVersion : controller.version,
-          });
-        } else if (controller.name === "meshsync" && controller.status.includes("ENABLED")) {
-          self.setState({ meshSyncInstalled : true,
-            meshSyncVersion : controller.version,
-            meshSyncState : controller.status
-          });
+  const columns = [
+    {
+      name : "contexts",
+      label : "Contexts",
+      options : {
+        filter : true,
+        sort : true,
+        searchable : true,
+        customBodyRender : (value, tableMeta, updateValue) => {
+          console.log("v", value, "tm", tableMeta, "uv", updateValue);
+          return (
+            <Tooltip title={`Server: ${tableMeta.rowData[2]}`}>
+              <Chip
+                label={data[tableMeta.rowIndex].name}
+                onDelete={() => handleConfigDelete(data[tableMeta.rowIndex].id, tableMeta.rowIndex)}
+                onClick={() => handleKubernetesClick(data[tableMeta.rowIndex].id, tableMeta.rowIndex)}
+                icon={<img src="/static/img/kubernetes.svg" className={classes.icon} />}
+                variant="outlined"
+                data-cy="chipContextName"
+              />
+            </Tooltip>
+          )
         }
-      });
-      self.setState({ operatorInstalled : true,
-        operatorSwitch : true,
-        operatorVersion : res.operator?.version, });
-      return true;
-    }
-
-    if (res.operator?.status === "DISABLED") self.setState({ operatorProcessing : false })
-
-    if (res.operator?.status === "PROCESSING") {
-      console.log("setting to processing");
-      self.setState({ operatorProcessing : true });
-    }
-
-
-    self.setState({
-      operatorInstalled : false,
-      NATSState : "UNKNOWN",
-      meshSyncInstalled : false,
-      meshSyncState : "N/A",
-      operatorSwitch : false,
-      operatorVersion : "N/A",
-      meshSyncVersion : "N/A",
-      NATSVersion : "N/A",
-    });
-
-    return false;
-  };
-
-  handleOperatorSwitch = () => {
-    const self = this;
-    const variables = { status : `${!self.state.operatorSwitch
-      ? "ENABLED"
-      : "DISABLED"}`, };
-    self.props.updateProgress({ showProgress : true });
-
-    changeOperatorState((response, errors) => {
-      self.props.updateProgress({ showProgress : false });
-      if (errors !== undefined) {
-        self.handleError("Unable to install operator");
       }
-      self.props.enqueueSnackbar("Operator " + response.operatorStatus.toLowerCase(), { variant : "success",
-        autoHideDuration : 2000,
-        action : (key) => (
-          <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
-            <CloseIcon />
-          </IconButton>
-        ), });
-      self.setState((state) => ({ operatorSwitch : !state.operatorSwitch }));
-    }, variables);
-  };
-
-  selectcontext = () => {
-    const self = this;
-    return (event) => {
-      self.setState(() => {
-        return { selectContext : event.target.value }
-      })
-    };
-  };
-
-  handleChange = (name) => {
-    const self = this;
-    return (event) => {
-      if (name === "inClusterConfigForm") {
-        self.setState({ [name] : event.target.checked, ts : new Date() });
-        return;
+    },
+    {
+      name : "deployment_type",
+      label : "Type of Deployment",
+      options : {
+        filter : true,
+        sort : true,
+        searchable : true,
       }
-      if (name === "k8sfile") {
-        if (event.target.value !== "") {
-          self.setState({ k8sfileError : false });
-        }
-
-        uploadK8sConfig().
-          then((name) => {
-            this.handleSuccess("successfully uploaded kubernetes config")
-            this.setState({ fileName : name });
-            fetchAllContexts(25)
-              .then(res => this.setState({ contexts : res.contexts }))
-              .catch(this.handleError("failed to get contexts"))
-          }).
-          catch(err => {
-            self.setState({ k8sfileError : true });
-            this.handleError("failed to upload kubernetes config")(err)
-          })
+    },
+    {
+      name : "location",
+      label : "Location",
+      options : {
+        filter : true,
+        sort : true,
+        searchable : true,
       }
-      if (name === "context") {
-        this.setState(() => {
-          return { contextName : this.state.selectContext }
-        })
-        let contextid;
-        this.state.contexts.forEach(element => {
-          if (element.name===this.state.selectContext){
-            contextid=element.id;
-          }
-        });
-        changeContext(contextid).
-          then(() => {
-            this.handleSuccess("successfully changed kubernetes current context")
-            fetchAllContexts(25)
-              .then(res => this.setState({ contexts : res.contexts }))
-              .catch(this.handleError("failed to get contexts"))
-          }).
-          catch(this.handleError("failed to change kubernetes current context"))
+    },
+    {
+      name : "last_discovery",
+      label : "Last Discovery",
+      options : {
+        filter : true,
+        sort : true,
+        searchable : true,
       }
-    };
-  };
-
-  handleKubernetesClick = (context) => {
-    this.props.updateProgress({ showProgress : true });
-    const self = this;
-    dataFetch(
-      "/api/system/kubernetes/ping?context=" + context,
-      { credentials : "same-origin" },
-      (result) => {
-        this.props.updateProgress({ showProgress : false });
-        if (typeof result !== "undefined") {
-          this.props.enqueueSnackbar("Kubernetes was successfully pinged!", {
-            variant : "success",
-            "data-cy" : "k8sSuccessSnackbar",
-            autoHideDuration : 2000,
-            action : (key) => (
-              <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
-                <CloseIcon />
+    },
+    {
+      name : "Actions",
+      options : {
+        filter : true,
+        sort : true,
+        searchable : true,
+        customBodyRender : (value, tableMeta) => {
+          return (
+            <div>
+              <IconButton
+                aria-label="more"
+                id="long-button"
+                aria-controls={showMenu[tableMeta.rowIndex] ? 'long-menu' : undefined}
+                aria-expanded={showMenu[tableMeta.rowIndex] ? 'true' : undefined}
+                aria-haspopup="true"
+                onClick={(e) => handleMenuOpen(e, tableMeta.rowIndex)}
+              >
+                <MoreVertIcon />
               </IconButton>
-            ),
-          });
-        }
+              <Menu
+                id="long-menu"
+                MenuListProps={{
+                  'aria-labelledby' : 'long-button',
+                }}
+                anchorEl={anchorEl}
+                open={showMenu[tableMeta.rowIndex]}
+                onClose={() => handleMenuClose(tableMeta.rowIndex)}
+                PaperProps={{
+                  style : {
+                    maxHeight : 48 * 4.5,
+                    width : '20ch',
+                  },
+                }}
+              >
+                <MenuItem onClick={() => {
+                  console.log("FMS"); handleMenuClose(tableMeta.rowIndex)
+                }}>
+                  FlushMeshSync
+                </MenuItem>
+                <MenuItem>
+                  <Switch
+                    checked={operatorSwitch}
+                    onClick={handleOperatorSwitch}
+                    disabled={operatorProcessing}
+                    name="OperatorSwitch"
+                    color="primary"
+                  />
+                    Operator
+                </MenuItem>
+              </Menu>
+            </div>
+          )
+        },
       },
-      self.handleError("Kubernetes config could not be validated")
-    );
-  };
 
-  handleOperatorClick = () => {
-    this.props.updateProgress({ showProgress : true });
-    const self = this;
+    }
+  ]
+
+  const options = {
+    print : false,
+    download : false,
+    expandableRows : true,
+    expandableRowsOnClick : false,
+    onRowsDelete : (td) => {
+      td.data.forEach((item) => {
+        console.log("D", item);
+        handleConfigDelete(data[item.index].id)
+      })
+    },
+    renderExpandableRow : (rowData, rowMetaData) => {
+      return (
+        <NoSsr>
+          <TableCell colSpan={6}>
+            <TableContainer>
+              <Table>
+                {/* <TableRow> */}
+                <TableCell className={classes.configBoxContainer}>
+                  <Paper >
+                    <div>
+                      <Grid container spacing={1} >
+                        <Grid item xs={12} md={5} className={classes.operationButton}>
+                          <List>
+                            <ListItem>
+                              <Tooltip title={`Server: ${rowData[2].server}`}
+                              >
+                                <Chip
+                                  label={data[rowMetaData.rowIndex].name}
+                                  onClick={() => handleKubernetesClick(data[rowMetaData.rowIndex].id, rowMetaData.rowIndex)}
+                                  icon={<img src="/static/img/kubernetes.svg" className={classes.icon} />}
+                                  variant="outlined"
+                                  data-cy="chipContextName"
+                                />
+                              </Tooltip>
+                            </ListItem>
+                          </List>
+                        </Grid>
+                      </Grid>
+                      <Grid container spacing={1} className={classes.contentContainer}>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText primary="Name" secondary={contexts[rowMetaData.rowIndex].name}/>
+                            </ListItem>
+                            <ListItem>
+                              <ListItemText primary="K8s Version" secondary={k8sVersion} />
+                            </ListItem>
+                          </List>
+                        </Grid>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText primary="Created At" secondary={
+                                contexts[rowMetaData.rowIndex].created_at
+                              }/>
+                            </ListItem>
+                            <ListItem>
+                              <ListItemText primary="UpdatedAt" secondary={
+                                contexts[rowMetaData.rowIndex].updated_at
+                              } />
+                            </ListItem>
+                          </List>
+                        </Grid>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText className={classes.text} primary="Server" secondary={
+                                contexts[rowMetaData.rowIndex].server
+                              }/>
+                            </ListItem>
+                          </List>
+                        </Grid>
+                      </Grid>
+                    </div>
+                  </Paper>
+                </TableCell>
+                <TableCell className={classes.configBoxContainer}>
+                  <Paper >
+                    <div>
+                      <Grid container spacing={1} >
+                        <Grid item xs={12} md={4} className={classes.operationButton}>
+                          <List>
+                            <ListItem>
+                              <Tooltip
+                                title={operatorInstalled
+                                  ? `Version: ${operatorVersion}`
+                                  : "Not Available"}
+                                aria-label="meshSync"
+                              >
+                                <Chip
+                                  // label={inClusterConfig?'Using In Cluster Config': contextName + (configuredServer?' - ' + configuredServer:'')}
+                                  label={"Operator"}
+                                  // onDelete={handleReconfigure}
+                                  onClick={handleOperatorClick}
+                                  icon={<img src="/static/img/meshery-operator.svg" className={classes.icon} />}
+                                  variant="outlined"
+                                  data-cy="chipOperator"
+                                />
+                              </Tooltip>
+                            </ListItem>
+                          </List>
+                        </Grid>
+                        {operatorInstalled &&
+                            <>
+                              <Grid item xs={12} md={4}>
+                                <List>
+                                  <ListItem>
+                                    <Tooltip
+                                      title={meshSyncInstalled ? `Redeploy MeshSync` : "Not Available"}
+                                      aria-label="meshSync"
+                                    >
+                                      <Chip
+                                        label={"MeshSync"}
+                                        onClick={handleMeshSyncClick}
+                                        icon={<img src="/static/img/meshsync.svg" className={classes.icon} />}
+                                        variant="outlined"
+                                        data-cy="chipMeshSync"
+                                      />
+                                    </Tooltip>
+                                  </ListItem>
+                                </List>
+                              </Grid>
+                              <Grid item xs={12} md={4}>
+                                <List>
+                                  <ListItem>
+                                    <Tooltip
+                                      title={NATSState === "CONNECTED" ? `Reconnect NATS` : "Not Available"}
+                                      aria-label="nats"
+                                    >
+                                      <Chip
+                                        label={"NATS"}
+                                        onClick={handleNATSClick}
+                                        icon={<img src="/static/img/nats-icon-color.svg" className={classes.icon} />}
+                                        variant="outlined"
+                                        data-cy="chipNATS"
+                                      />
+                                    </Tooltip>
+                                  </ListItem>
+                                </List>
+                              </Grid>
+                            </>
+                        }
+                      </Grid>
+                      <Grid container spacing={1} className={classes.contentContainer}>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText primary="Operator State" secondary={operatorInstalled
+                                ? "Active"
+                                : "Disabled"} />
+                            </ListItem>
+                            <ListItem>
+                              <ListItemText primary="Operator Version" secondary={operatorVersion} />
+                            </ListItem>
+                          </List>
+                        </Grid>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText primary="MeshSync State" secondary={meshSyncInstalled
+                                ? meshSyncState
+                                : "Disabled"} />
+                            </ListItem>
+                            <ListItem>
+                              <ListItemText primary="MeshSync Version" secondary={meshSyncVersion} />
+                            </ListItem>
+                          </List>
+                        </Grid>
+                        <Grid item xs={12} md={5}>
+                          <List>
+                            <ListItem>
+                              <ListItemText primary="NATS State" secondary={NATSState} />
+                            </ListItem>
+                            <ListItem>
+                              <ListItemText primary="NATS Version" secondary={NATSVersion} />
+                            </ListItem>
+                          </List>
+                        </Grid>
+                      </Grid>
+                    </div>
+                  </Paper>
+                </TableCell>
+                {/* </TableRow> */}
+              </Table>
+            </TableContainer>
+          </TableCell>
+        </NoSsr>
+      )
+    }
+  }
+
+  const handleClick = async () => {
+    const modal = ref.current;
+    let response = await modal.show({
+      title : "Add Kuberneted Cluster(s)",
+      subtitle :
+      <>
+        <div>
+          <Typography variant="h6">
+          Upload your kubeconfig
+          </Typography>
+          <Typography variant="body2">
+          commonly found at ~/.kube/config
+          </Typography>
+          <FormGroup>
+            <input
+              id="k8sfile"
+              type="file"
+              value={k8sfileElementVal}
+              onChange={handleChange}
+              className={classes.fileInputStyle}
+            />
+            <TextField
+              id="k8sfileLabelText"
+              name="k8sfileLabelText"
+              className={classes.fileLabelText}
+              label="Upload kubeconfig"
+              variant="outlined"
+              fullWidth
+              onClick={() => {
+                document.querySelector("#k8sfile")?.click(); setOpen(true);
+              }}
+              margin="normal"
+              InputProps={{ readOnly : true,
+                endAdornment : (
+                  <InputAdornment position="end">
+                    <CloudUploadIcon />
+                  </InputAdornment>
+                ), }}
+            />
+          </FormGroup>
+          {/* <Dialog
+            open={open}
+            onClose={handleClose}
+            aria-labelledby="alert-dialog-title"
+            aria-describedby="alert-dialog-description"
+            className={classes.dialogBox}
+          >
+            <DialogContent>
+              <DialogContentText className={ classes.subtitle }>
+                <Typography>
+                    Available Contexts
+                </Typography>
+                {
+                  toUploadContexts.map((ctx) => (
+                    <Chip
+                      label={ctx.name}
+                      // onDelete={() => handleReconfigure(ctx.id)}
+                      onClick={() => handleKubernetesClick(data[tableMeta.rowIndex].id, tableMeta.rowIndex)}
+                      icon={<img src="/static/img/kubernetes.svg" className={classes.icon} />}
+                      variant="outlined"
+                      data-cy="chipContextName"
+                    />
+                  ))
+                }
+              </DialogContentText>
+            </DialogContent>
+          </Dialog> */}
+        </div>
+      </>,
+      options : ["Cancel", "Upload"]
+    })
+    if (response === "Upload") {
+      uploadK8SConfig().then(() => {
+        handleSuccess("successfully uploaded kubernetes config");
+        fetchAllContexts(25)
+          .then(res => setContexts(res.contexts))
+          .catch(handleError("failed to get contexts"))
+      }).
+        catch(err => {
+          handleError("failed to upload kubernetes config")(err)
+        })
+    }
+  }
+
+  const handleOperatorClick = () => {
+    updateProgress({ showProgress : true });
     fetchMesheryOperatorStatus().subscribe({ next : (res) => {
       console.log(res);
-      let state = self.setOperatorState(res);
-      self.props.updateProgress({ showProgress : false });
+      let state = setOperatorState(res);
+      updateProgress({ showProgress : false });
       if (state == true) {
-        this.props.enqueueSnackbar("Operator was successfully pinged!", { variant : "success",
+        enqueueSnackbar("Operator was successfully pinged!", { variant : "success",
           autoHideDuration : 2000,
           action : (key) => (
-            <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
+            <IconButton key="close" aria-label="Close" color="inherit" onClick={() => closeSnackbar(key)}>
               <CloseIcon />
             </IconButton>
           ), });
       } else {
-        self.handleError("Operator could not be reached")("Operator is disabled");
+        handleError("Operator could not be reached")("Operator is disabled");
       }
     },
-    error : self.handleError("Operator could not be pinged"), });
+    error : handleError("Operator could not be pinged"), });
   };
 
-handleNATSClick = () => {
-  this.props.updateProgress({ showProgress : true });
-  const self = this;
+  const handleNATSClick = () => {
+    updateProgress({ showProgress : true });
+    NatsStatusQuery().subscribe({
+      next : (res) => {
+        updateProgress({ showProgress : false });
+        if (res.controller.name === "broker" && res.controller.status.includes("CONNECTED")) {
+          let runningEndpoint = res.controller.status.substring("CONNECTED".length)
+          enqueueSnackbar(`Broker was successfully pinged. Running at ${runningEndpoint}`, {
+            variant : "success",
+            action : (key) => (
+              <IconButton key="close" aria-label="close" color="inherit" onClick={() => closesnackbar(key)}>
+                <CloseIcon />
+              </IconButton>
+            ),
+            autohideduration : 2000,
+          })
+        } else {
+          handleError("Meshery Broker could not be reached")("Meshery Server is not connected to Meshery Broker");
+        }
+        setNATSState(res.controller.status.length !== 0 ? res.controller.status : "UNKNOWN");
+        setNATSVersion(res.controller.version);
+      },
+      error : handleError("NATS status could not be retrieved"), });
 
-  NatsStatusQuery().subscribe({
-    next : (res) => {
-      self.props.updateProgress({ showProgress : false });
-      if (res.controller.name === "broker" && res.controller.status.includes("CONNECTED")) {
-        let runningEndpoint = res.controller.status.substring("CONNECTED".length)
-        this.props.enqueueSnackbar(`Broker was successfully pinged. Running at ${runningEndpoint}`, {
-          variant : "success",
-          action : (key) => (
-            <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
-              <CloseIcon />
-            </IconButton>
-          ),
-          autohideduration : 2000,
-        })
-      } else {
-        self.handleError("Meshery Broker could not be reached")("Meshery Server is not connected to Meshery Broker");
-      }
-      self.setState({
-        NATSState : res.controller.status.length !== 0 ? res.controller.status : "UNKNOWN",
-        NATSVersion : res.controller.version,
-      });
-    },
-    error : self.handleError("NATS status could not be retrieved"), });
-
-  // connectToNats().subscribe({
-  //   next : (res) => {
-  //     if (res.connectToNats === "PROCESSING") {
-  //       this.props.updateProgress({ showProgress : false });
-  //       this.props.enqueueSnackbar(`Reconnecting to NATS...`, {
-  //         variant : "info",
-  //         action : (key) => (
-  //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
-  //             <CloseIcon />
-  //           </IconButton>
-  //         ),
-  //         autohideduration : 7000,
-  //       })
-  //     }
-  //     if (res.connectToNats === "CONNECTED") {
-  //       this.props.updateProgress({ showProgress : false });
-  //       this.props.enqueueSnackbar(`Successfully connected to NATS`, {
-  //         variant : "success",
-  //         action : (key) => (
-  //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
-  //             <CloseIcon />
-  //           </IconButton>
-  //         ),
-  //         autohideduration : 7000,
-  //       })
-  //     }
-
-  //   },
-  //   error : self.handleError("Failed to request reconnection with NATS"),
-  // });
-
-};
-
-  handleMeshSyncClick = () => {
-    this.props.updateProgress({ showProgress : true });
-    const self = this;
-
-    MeshsyncStatusQuery().subscribe({ next : (res) => {
-      self.props.updateProgress({ showProgress : false });
-      if (res.controller.name === "meshsync" && res.controller.status.includes("ENABLED")) {
-        self.setState({
-          meshSyncInstalled : true,
-          meshSyncVersion : res.controller.version,
-          meshSyncState : res.controller.status,
-        });
-        let publishEndpoint = res.controller.status.substring("ENABLED".length)
-        this.props.enqueueSnackbar(`MeshSync was successfully pinged. Publishing to ${publishEndpoint} `, {
-          variant : "success",
-          action : (key) => (
-            <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
-              <CloseIcon />
-            </IconButton>
-          ),
-          autohideduration : 2000,
-        })
-      } else {
-        self.handleError("MeshSync could not be reached")("MeshSync is unavailable");
-        self.setState({
-          meshSyncInstalled : false,
-          meshSyncVersion : "",
-          meshSyncState : res.controller.status
-        });
-      }
-    },
-    error : self.handleError("MeshSync status could not be retrieved"), });
-
-    // deployMeshSync().subscribe({
+    // connectToNats().subscribe({
     //   next : (res) => {
-    //     if (res.deployMeshsync === "PROCESSING") {
-    //       this.props.updateProgress({ showProgress : false });
-    //       this.props.enqueueSnackbar(`MeshSync deployment in progress`, {
+    //     if (res.connectToNats === "PROCESSING") {
+    //       updateProgress({ showProgress : false });
+    //       enqueueSnackbar(`Reconnecting to NATS...`, {
     //         variant : "info",
     //         action : (key) => (
-    //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
+    //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => closesnackbar(key)}>
+    //             <CloseIcon />
+    //           </IconButton>
+    //         ),
+    //         autohideduration : 7000,
+    //       })
+    //     }
+    //     if (res.connectToNats === "CONNECTED") {
+    //       updateProgress({ showProgress : false });
+    //       enqueueSnackbar(`Successfully connected to NATS`, {
+    //         variant : "success",
+    //         action : (key) => (
+    //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => closesnackbar(key)}>
     //             <CloseIcon />
     //           </IconButton>
     //         ),
@@ -542,518 +807,86 @@ handleNATSClick = () => {
     //     }
 
     //   },
-    //   error : self.handleError("Failed to request Meshsync redeployment"),
+    //   error : handleError("Failed to request reconnection with NATS"),
+    // });
+
+  };
+
+  const handleMeshSyncClick = () => {
+    updateProgress({ showProgress : true });
+    MeshsyncStatusQuery().subscribe({ next : (res) => {
+      updateProgress({ showProgress : false });
+      if (res.controller.name === "meshsync" && res.controller.status.includes("ENABLED")) {
+        setMeshSyncInstalled(true);
+        setMeshSyncVersion(res.controller.version);
+        setMeshSyncState(res.controller.status);
+        let publishEndpoint = res.controller.status.substring("ENABLED".length)
+        enqueueSnackbar(`MeshSync was successfully pinged. Publishing to ${publishEndpoint} `, {
+          variant : "success",
+          action : (key) => (
+            <IconButton key="close" aria-label="close" color="inherit" onClick={() => closesnackbar(key)}>
+              <CloseIcon />
+            </IconButton>
+          ),
+          autohideduration : 2000,
+        })
+      } else {
+        handleError("MeshSync could not be reached")("MeshSync is unavailable");
+        setMeshSyncInstalled(false);
+        setMeshSyncVersion("");
+        setMeshSyncState(res.controller.status);
+      }
+    },
+    error : handleError("MeshSync status could not be retrieved"), });
+
+    // deployMeshSync().subscribe({
+    //   next : (res) => {
+    //     if (res.deployMeshsync === "PROCESSING") {
+    //       updateProgress({ showProgress : false });
+    //       enqueueSnackbar(`MeshSync deployment in progress`, {
+    //         variant : "info",
+    //         action : (key) => (
+    //           <IconButton key="close" aria-label="close" color="inherit" onClick={() => closesnackbar(key)}>
+    //             <CloseIcon />
+    //           </IconButton>
+    //         ),
+    //         autohideduration : 7000,
+    //       })
+    //     }
+
+    //   },
+    //   error : handleError("Failed to request Meshsync redeployment"),
     // });
   };
 
-  handleReset = () => {
-    this.props.updateProgress({ showProgress : true });
-    const self = this;
-    resetDatabase({
-      selector : {
-        clearDB : "true",
-        ReSync : "true", // True - For Hard reset, False otherwise
-        hardReset : "true",
-      },
-      // For now by default set to hard reset.
-      // TODO: User should be able to select soft or hard reset when "Reset Daatabase" btn is clicked.
-    }).subscribe({
-      next : (res) => {
-        self.props.updateProgress({ showProgress : false });
-        if (res.resetStatus === "PROCESSING") {
-          this.props.enqueueSnackbar(`Database reset successful.`, {
-            variant : "success",
-            action : (key) => (
-              <IconButton key="close" aria-label="close" color="inherit" onClick={() => self.props.closesnackbar(key)}>
-                <CloseIcon />
-              </IconButton>
-            ),
-            autohideduration : 2000,
-          })
-        }
-      },
-      error : self.handleError("Database is not reachable, try restarting server.")
-    });
-  }
+  return (
+    <>
+      <Button
+        type="submit"
+        variant="contained"
+        color="primary"
+        size="large"
+        onClick={handleClick}
+        className={classes.button}
+        data-cy="btnResetDatabase"
+      >
+      Add Cluster
+      </Button>
 
-  handleError = (msg) => (error) => {
-    this.props.updateProgress({ showProgress : false });
-    const self = this;
-    this.props.enqueueSnackbar(`${msg}: ${error}`, { variant : "error", preventDuplicate : true,
-      action : (key) => (
-        <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
-          <CloseIcon />
-        </IconButton>
-      ),
-      autoHideDuration : 7000, });
-  };
-
-  handleSuccess = msg => {
-    this.props.updateProgress({ showProgress : false });
-    const self = this;
-    this.props.enqueueSnackbar(msg, {
-      variant : "success",
-      action : (key) => (
-        <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
-          <CloseIcon />
-        </IconButton>
-      ),
-      autoHideDuration : 7000,
-    });
-  }
-
-  handleReconfigure = (id) => {
-    dataFetch(
-      "/api/system/kubernetes/contexts/" + id,
-      {
-        credentials : "same-origin",
-        method : "DELETE"
-      },
-      (result) => {
-        this.props.updateProgress({ showProgress : false });
-        if (result) {
-          const modal = this.ref.current;
-          const self = this;
-
-          if (self.state.operatorSwitch) {
-            setTimeout(async () => {
-              let response = await modal.show({
-                title : "Remove Meshery Operator from this cluster?",
-                subtitle : "Meshery is now disconnected from your Kubernetes cluster. Do you want to remove the Meshery Operator from your cluster as well?",
-                options : ["yes", "no"]
-              });
-
-              if (response === "yes") {
-                const variables = { status : "DISABLED", };
-                self.props.updateProgress({ showProgress : true });
-
-                changeOperatorState((response, errors) => {
-                  self.props.updateProgress({ showProgress : false });
-
-                  if (errors !== undefined) self.handleError("Operator action failed");
-
-                  self.props.enqueueSnackbar(
-                    "Operator " + response.operatorStatus.toLowerCase(),
-                    {
-                      variant : "success",
-                      autoHideDuration : 2000,
-                      action : (key) => (
-                        <IconButton
-                          key="close"
-                          aria-label="Close"
-                          color="inherit"
-                          onClick={() => self.props.closeSnackbar(key)}
-                        >
-                          <CloseIcon />
-                        </IconButton>
-                      ),
-                    }
-                  );
-
-                  self.setState((state) => ({ operatorSwitch : !state.operatorSwitch }));
-                }, variables);
-              }
-            }, 100);
-          }
-
-          // Remove the context from the state
-          self.setState(state => {
-            return { contexts : state?.contexts?.filter(ctx => ctx.id !== id) }
-          })
-
-          // Display the success message
-          this.props.enqueueSnackbar(
-            "Kubernetes context was successfully removed!",
-            {
-              variant : "success",
-              autoHideDuration : 2000,
-              action : (key) => (
-                <IconButton key="close" aria-label="Close" color="inherit" onClick={() => self.props.closeSnackbar(key)}>
-                  <CloseIcon />
-                </IconButton>
-              ),
-            }
-          );
-        }
-      },
-      this.handleError("failed to delete kubernetes context")
-    )
-  };
-
-  configureTemplate = () => {
-    const { classes } = this.props;
-    const {
-      inClusterConfig,
-      contextName,
-      clusterConfigured,
-      configuredServer,
-      operatorInstalled,
-      operatorVersion,
-      meshSyncInstalled,
-      meshSyncVersion,
-      meshSyncState,
-      NATSState,
-      NATSVersion,
-      operatorSwitch,
-      contexts,
-    } = this.state;
-
-    let showConfigured = <></>;
-    const self = this;
-    if (clusterConfigured) {
-      const lst = (
-        <List>
-          <ListItem>
-            <ListItemText
-              primary="Context Name"
-              secondary={inClusterConfig
-                ? "Using In Cluster Config"
-                : contextName}
-              data-cy="itemListContextName"
-            />
-          </ListItem>
-          <ListItem>
-            <ListItemText
-              primary="Server Name"
-              secondary={inClusterConfig
-                ? "In Cluster Server"
-                : configuredServer || ""}
-              data-cy="itemListServerName"
-            />
-          </ListItem>
-        </List>
-      );
-      showConfigured = (
-        <div>
-          {contexts?.map(ctx => (
-            <Tooltip title={`Server: ${ctx.server}`}>
-              <Chip
-                label={ctx?.name}
-                onDelete={() => self.handleReconfigure(ctx.id)}
-                onClick={() => self.handleKubernetesClick(ctx.id)}
-                icon={<img src="/static/img/kubernetes.svg" className={classes.icon} />}
-                variant="outlined"
-                data-cy="chipContextName"
-              />
-            </Tooltip>
-          ))}
-          {lst}
-        </div>
-      );
-    }
-    if (!clusterConfigured) {
-      const lst = (
-        <List>
-          <ListItem>
-            <ListItemText primary="Context Name" secondary="Not Configured" />
-          </ListItem>
-          <ListItem>
-            <ListItemText primary="Server Name" secondary="Not Configured" />
-          </ListItem>
-        </List>
-      );
-      showConfigured = <div>{lst}</div>;
-    }
-
-    const operator = (
-      <React.Fragment>
-        <div>
-          <Grid container spacing={1} >
-            <Grid item xs={12} md={4} className={classes.operationButton}>
-              <List>
-                <ListItem>
-                  <Tooltip
-                    title={operatorInstalled
-                      ? `Version: ${operatorVersion}`
-                      : "Not Available"}
-                    aria-label="meshSync"
-                  >
-                    <Chip
-                      // label={inClusterConfig?'Using In Cluster Config': contextName + (configuredServer?' - ' + configuredServer:'')}
-                      label={"Operator"}
-                      // onDelete={self.handleReconfigure}
-                      onClick={self.handleOperatorClick}
-                      icon={<img src="/static/img/meshery-operator.svg" className={classes.icon} />}
-                      variant="outlined"
-                      data-cy="chipOperator"
-                    />
-                  </Tooltip>
-                </ListItem>
-              </List>
-            </Grid>
-            {operatorInstalled &&
-            <>
-              <Grid item xs={12} md={4}>
-                <List>
-                  <ListItem>
-                    <Tooltip
-                      title={meshSyncInstalled ? `Redeploy MeshSync` : "Not Available"}
-                      aria-label="meshSync"
-                    >
-                      <Chip
-                        label={"MeshSync"}
-                        onClick={self.handleMeshSyncClick}
-                        icon={<img src="/static/img/meshsync.svg" className={classes.icon} />}
-                        variant="outlined"
-                        data-cy="chipMeshSync"
-                      />
-                    </Tooltip>
-                  </ListItem>
-                </List>
-              </Grid>
-              <Grid item xs={12} md={4}>
-                <List>
-                  <ListItem>
-                    <Tooltip
-                      title={NATSState === "CONNECTED" ? `Reconnect NATS` : "Not Available"}
-                      aria-label="nats"
-                    >
-                      <Chip
-                        label={"NATS"}
-                        onClick={self.handleNATSClick}
-                        icon={<img src="/static/img/nats-icon-color.svg" className={classes.icon} />}
-                        variant="outlined"
-                        data-cy="chipNATS"
-                      />
-                    </Tooltip>
-                  </ListItem>
-                </List>
-              </Grid>
-            </>
-            }
-          </Grid>
-          <Grid container spacing={1}>
-            <Grid item xs={12} md={4}>
-              <List>
-                <ListItem>
-                  <ListItemText primary="Operator State" secondary={operatorInstalled
-                    ? "Active"
-                    : "Disabled"} />
-                </ListItem>
-                <ListItem>
-                  <ListItemText primary="Operator Version" secondary={operatorVersion} />
-                </ListItem>
-              </List>
-            </Grid>
-            <Grid item xs={12} md={4}>
-              <List>
-                <ListItem>
-                  <ListItemText primary="MeshSync State" secondary={meshSyncInstalled
-                    ? meshSyncState
-                    : "Disabled"} />
-                </ListItem>
-                <ListItem>
-                  <ListItemText primary="MeshSync Version" secondary={meshSyncVersion} />
-                </ListItem>
-              </List>
-            </Grid>
-            <Grid item xs={12} md={4}>
-              <List>
-                <ListItem>
-                  <ListItemText primary="NATS State" secondary={NATSState} />
-                </ListItem>
-                <ListItem>
-                  <ListItemText primary="NATS Version" secondary={NATSVersion} />
-                </ListItem>
-              </List>
-            </Grid>
-          </Grid>
-        </div>
-        <div className={classes.grey}>
-          <FormGroup>
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={operatorSwitch}
-                  onClick={self.handleOperatorSwitch}
-                  disabled={self.state.operatorProcessing}
-                  name="OperatorSwitch"
-                  color="primary"
-                />
-              }
-              label="Meshery Operator"
-            />
-            {self.state.operatorProcessing && <CircularProgress />}
-          </FormGroup>
-        </div>
-        <div className={classes.buttonsCluster}>
-          <Button
-            type="submit"
-            variant="contained"
-            color="primary"
-            size="large"
-            onClick={() => this.handleReset()}
-            className={classes.button}
-            data-cy="btnResetDatabase"
-          >
-              Reset Database
-          </Button>
-        </div>
-      </React.Fragment>
-    );
-
-    if (this.props.tabs == 0) return this.meshOut(showConfigured, operator);
-
-    return this.meshIn(showConfigured, operator);
-  };
-
-  meshOut = (showConfigured, operator) => {
-    const { classes } = this.props;
-    const {
-      k8sfileElementVal, contextName, contexts, fileName, selectContext
-    } = this.state;
-
-    return (
-      <NoSsr>
-        <PromptComponent ref={this.ref} />
-        <div className={classes.clusterConfiguratorWrapper}>
-          <Grid container spacing={5} className={classes.contentContainer}>
-            <Grid item spacing={1} xs={12} md={6} className={classes.configBoxContainer}>
-              <div className={classes.heading}>
-                <h4>Cluster Configuration</h4>
-              </div>
-              <Paper className={classes.paper}>
-                <div>{showConfigured}</div>
-                <div className={classes.grey}>
-                  <FormGroup>
-                    <input
-                      id="k8sfile"
-                      type="file"
-                      value={k8sfileElementVal}
-                      onChange={this.handleChange("k8sfile")}
-                      className={classes.fileInputStyle}
-                    />
-                    <TextField
-                      id="k8sfileLabelText"
-                      name="k8sfileLabelText"
-                      className={classes.fileLabelText}
-                      label="Upload kubeconfig"
-                      variant="outlined"
-                      fullWidth
-                      value={fileName}
-                      onClick={() => document.querySelector("#k8sfile")?.click()}
-                      margin="normal"
-                      InputProps={{ readOnly : true,
-                        endAdornment : (
-                          <InputAdornment position="end">
-                            <CloudUploadIcon />
-                          </InputAdornment>
-                        ), }}
-                    />
-                  </FormGroup>
-                  <form onSubmit={this.handleChange("context")} >
-                    <Box sx={{ minWidth : 120 }}>
-                      <FormControl variant="outlined" fullWidth>
-                        <InputLabel id="select-label">Context Name</InputLabel>
-                        <Select
-                          id="contextName"
-                          name="contextName"
-                          labelId="select-label"
-                          label="Context Name"
-                          value={selectContext || contextName }
-                          onChange={this.selectcontext()}
-                        >
-                          {contexts?.map((ct) => (
-                            <MenuItem key={`ct_---_${ct.name}`} value= {ct.name}>
-                              {ct.name}
-                              {ct.is_current_context
-                                ? " (Current Context)"
-                                : ""}
-                            </MenuItem>
-                          ))}
-                        </Select>
-                      </FormControl>
-                    </Box>
-                    <div style={{ display : 'flex', justifyContent : 'center' }}>
-                      <Button
-                        type="submit"
-                        value="Submit"
-                        variant="contained"
-                        color="primary"
-                        size="large"
-                        style={{ margin : "0.5rem 0.5rem", whiteSpace : "nowrap" }}
-                      >
-                      Change Context
-                      </Button>
-                    </div>
-                  </form>
-                </div>
-              </Paper>
-            </Grid>
-
-            <Grid item spacing={1} xs={12} md={6} className={classes.configBoxContainer}>
-              <div className={classes.heading}>
-                <h4>Operator Configuration</h4>
-              </div>
-              <Paper className={classes.paper}>{operator}</Paper>
-            </Grid>
-          </Grid>
-        </div>
-      </NoSsr>
-    );
-  };
-
-  meshIn = (showConfigured, operator) => {
-    const { classes } = this.props;
-
-    return (
-      <NoSsr>
-        <PromptComponent ref={this.ref} />
-        <div className={classes.clusterConfiguratorWrapper}>
-          <Grid container spacing={5}>
-            <Grid item spacing={1} xs={12} md={6}>
-              <div className={classes.heading}>
-                <h4>Cluster Configuration</h4>
-              </div>
-              <Paper className={classes.paper}>
-                <div>{showConfigured}</div>
-                <div className={classes.grey}>
-                  <div className={classes.buttonsCluster}>
-                    <Button
-                      type="submit"
-                      variant="contained"
-                      color="primary"
-                      size="large"
-                      onClick={() => window.location.reload()}
-                      className={classes.button}
-                      data-cy="btnDiscoverCluster"
-                    >
-                      Discover Cluster
-                    </Button>
-                  </div>
-                </div>
-              </Paper>
-            </Grid>
-
-            <Grid item xs={12} md={6} spacing={1}>
-              <div className={classes.heading}>
-                <h4>Operator Configuration</h4>
-              </div>
-              <Paper className={classes.paper}>{operator}</Paper>
-            </Grid>
-          </Grid>
-        </div>
-      </NoSsr>
-    );
-  };
-
-  render() {
-
-    return this.configureTemplate();
-  }
+      <DataTable
+        columns = { columns }
+        data = { data }
+        options = { options }
+      />
+      <PromptComponent ref={ ref }/>
+    </>
+  )
 }
-
-MeshConfigComponent.propTypes = { classes : PropTypes.object.isRequired, };
-
-const mapDispatchToProps = (dispatch) => ({ updateK8SConfig : bindActionCreators(updateK8SConfig, dispatch),
-  updateProgress : bindActionCreators(updateProgress, dispatch), });
 const mapStateToProps = (state) => {
-  const k8sconfig = state.get("k8sConfig").toJS();
+  const k8sconfig = state.get('k8sConfig').toJS();
   return k8sconfig;
-};
+}
+const mapDispatchToProps = (dispatch) => ({ updateProgress : bindActionCreators(updateProgress, dispatch), });
 
-// @ts-ignore
-export default withStyles(styles)(
-  // @ts-ignore
-  connect(mapStateToProps, mapDispatchToProps)(withRouter(withSnackbar(MeshConfigComponent)))
-);
+export default withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(withSnackbar(MesherySettingsNew)));
+
